added server side location updates and avatar movement
also major fix to script loading issues utilizing jquery promise in
loadscripts.js

+++ EdenModels.py
@@ -16,6 +16,6 @@ class Pulse(ndb.Model):
     origin = ndb.StringProperty()
 
 class PulseLoc(Pulse):
-    xloc = ndb.IntegerProperty()
-    yloc = ndb.IntegerProperty()
-    zloc = ndb.IntegerProperty()
+    xloc = ndb.FloatProperty()
+    yloc = ndb.FloatProperty()
+    zloc = ndb.FloatProperty()
